two "new" statuses
We need two statuses for tracking how fresh is the document for a given
user.

(1) The first status "is_new" is set to "1" when the user receives
document for the first time. When the user opens the document, this flag
is set to "0" and never changes again. The only exception may happen
when the user desides to mark the document as "unread". When user marks
the document as "unread", "is_new" status is set back to "1".

(2) The second status "is_changed" is set to "1" whenever new comment or
any kind of change appears in the document. When the user opens the
document, this flag is set to "0". But it can be changed back to "1" if
any other change occures.

Other users can track only "is_new" status. They have no idea about
"is_changed" status.